refactor(backend): optimize other resource download backend code style, add a fallback to curseforge download url
