+++ src-tauri/src/instance/helpers/modpack/curseforge.rs
@@ -13,7 +13,7 @@ use zip::ZipArchive;
 use crate::{
   error::{SJMCLError, SJMCLResult},
   instance::models::misc::{InstanceError, ModLoaderType},
-  resource::helpers::curseforge::CurseForgeProject,
+  resource::helpers::curseforge_misc::CurseForgeProject,
   tasks::{download::DownloadParam, PTaskParam},
 };
