+++ src-tauri/src/resource/helpers/curseforge_misc.rs
@@ -1,6 +1,155 @@
 use lazy_static::lazy_static;
 use std::collections::HashMap;
 
+use crate::error::SJMCLResult;
+use crate::resource::models::{
+  OtherResourceDependency, OtherResourceFileInfo, OtherResourceInfo, OtherResourceSearchRes,
+  OtherResourceSource, OtherResourceVersionPack, ResourceError,
+};
+use serde::{Deserialize, Serialize};
+
+use super::misc::version_pack_sort;
+
+#[derive(Deserialize, Serialize, Debug, Clone)]
+#[serde(rename_all = "camelCase")]
+pub struct CurseForgeCategory {
+  pub name: String,
+}
+
+#[derive(Deserialize, Serialize, Debug, Clone)]
+#[serde(rename_all = "camelCase")]
+pub struct CurseForgeLink {
+  pub website_url: String,
+}
+
+#[derive(Deserialize, Serialize, Debug, Clone)]
+#[serde(rename_all = "camelCase")]
+pub struct CurseForgeLogo {
+  pub url: String,
+}
+
+pub fn default_logo() -> CurseForgeLogo {
+  CurseForgeLogo {
+    url: "".to_string(),
+  }
+}
+
+#[derive(Deserialize, Serialize, Debug, Clone)]
+#[serde(rename_all = "camelCase")]
+pub struct CurseForgeProject {
+  pub id: u32,
+  pub class_id: u32,
+  pub links: CurseForgeLink,
+  pub name: String,
+  pub summary: String,
+  pub categories: Vec<CurseForgeCategory>,
+  pub download_count: u32,
+  pub logo: Option<CurseForgeLogo>, // In some old projects, logo is null
+  pub date_modified: String,
+}
+
+#[derive(Deserialize, Serialize, Debug, Clone)]
+#[serde(rename_all = "camelCase")]
+pub struct CurseForgePagination {
+  pub index: u32,
+  pub page_size: u32,
+  pub total_count: u32,
+}
+
+#[derive(Deserialize, Serialize, Debug, Clone)]
+pub struct CurseForgeSearchRes {
+  pub data: Vec<CurseForgeProject>,
+  pub pagination: CurseForgePagination,
+}
+
+#[derive(Deserialize, Serialize, Debug, Clone)]
+pub struct CurseForgeFileHash {
+  pub value: String,
+  pub algo: u32,
+}
+
+#[derive(Deserialize, Serialize, Debug, Clone)]
+#[serde(rename_all = "camelCase")]
+pub struct CurseForgeFileDependency {
+  pub mod_id: u32,
+  pub relation_type: u32,
+}
+
+#[derive(Deserialize, Serialize, Debug, Clone)]
+#[serde(rename_all = "camelCase")]
+pub struct CurseForgeFileInfo {
+  pub mod_id: u32,
+  pub display_name: String,
+  pub file_name: String,
+  pub release_type: u32,
+  pub hashes: Vec<CurseForgeFileHash>,
+  pub file_date: String,
+  pub download_url: Option<String>,
+  pub download_count: u32,
+  pub game_versions: Vec<String>,
+  pub dependencies: Vec<CurseForgeFileDependency>,
+}
+
+#[derive(Deserialize, Serialize, Debug, Clone)]
+pub struct CurseForgeVersionPackSearchRes {
+  pub data: Vec<CurseForgeFileInfo>,
+  pub pagination: CurseForgePagination,
+}
+
+#[derive(Deserialize, Serialize, Debug, Clone)]
+#[serde(rename_all = "camelCase")]
+pub struct CurseForgeExactMatches {
+  pub file: CurseForgeFileInfo,
+}
+
+#[derive(Deserialize, Serialize, Debug, Clone)]
+#[serde(rename_all = "camelCase")]
+pub struct CurseForgeFingerprintData {
+  pub exact_matches: Vec<CurseForgeExactMatches>,
+}
+
+#[derive(Deserialize, Serialize, Debug, Clone)]
+#[serde(rename_all = "camelCase")]
+pub struct CurseForgeFingerprintRes {
+  pub data: CurseForgeFingerprintData,
+}
+
+#[derive(Deserialize, Serialize, Debug, Clone)]
+#[serde(rename_all = "camelCase")]
+pub struct CurseForgeGetProjectRes {
+  pub data: CurseForgeProject,
+}
+
+#[derive(Debug, Clone, Copy)]
+pub enum CurseForgeApiEndpoint {
+  Search,
+  ModFiles,
+  Fingerprints,
+  Project,
+}
+
+pub fn get_curseforge_api(
+  endpoint: CurseForgeApiEndpoint,
+  id: Option<&str>,
+) -> SJMCLResult<String> {
+  let base_url = "https://api.curseforge.com/v1";
+
+  let url_str = match endpoint {
+    CurseForgeApiEndpoint::Search => format!("{}/mods/search", base_url),
+    CurseForgeApiEndpoint::ModFiles => {
+      let mod_id = id.ok_or(ResourceError::ParseError)?;
+      format!("{}/mods/{}/files", base_url, mod_id)
+    }
+    CurseForgeApiEndpoint::Fingerprints => format!("{}/fingerprints/432", base_url),
+    CurseForgeApiEndpoint::Project => {
+      let mod_id = id.ok_or(ResourceError::ParseError)?;
+      format!("{}/mods/{}", base_url, mod_id)
+    }
+  };
+
+  Ok(url_str)
+}
+
 lazy_static! {
   static ref CATEGORY_MAP: HashMap<(String, u32), u32> = {
     let mut map = HashMap::new();
@@ -220,4 +369,141 @@ pub fn cvt_id_to_dependency_type(dependency_type: u32) -> String {
     5 => "incompatible".to_string(),
     _ => "include".to_string(),
   }
+}
+
+pub fn map_curseforge_file_info_to_other_resource_file_info(
+  cf_file: &CurseForgeFileInfo,
+  loader: Option<String>,
+) -> OtherResourceFileInfo {
+  let download_url = match &cf_file.download_url {
+    Some(url) => url.clone(),
+    None => {
+      // This addon is not allowed for distribution, and downloadUrl will be null.
+      // We try to find its download url.
+      format!(
+        "https://edge.forgecdn.net/files/{}/{}/{}",
+        cf_file.mod_id / 1000,
+        cf_file.mod_id % 1000,
+        cf_file.file_name
+      )
+    }
+  };
+
+  OtherResourceFileInfo {
+    resource_id: cf_file.mod_id.to_string(),
+    name: cf_file.display_name.clone(),
+    release_type: cvt_id_to_release_type(cf_file.release_type),
+    downloads: cf_file.download_count,
+    file_date: cf_file.file_date.clone(),
+    download_url,
+    sha1: cf_file
+      .hashes
+      .iter()
+      .find(|h| h.algo == 1)
+      .map_or("".to_string(), |h| h.value.clone()),
+    file_name: cf_file.file_name.clone(),
+    dependencies: cf_file
+      .dependencies
+      .iter()
+      .map(|dep| OtherResourceDependency {
+        resource_id: dep.mod_id.to_string(),
+        relation: cvt_id_to_dependency_type(dep.relation_type),
+      })
+      .collect(),
+    loader,
+  }
+}
+
+pub fn map_curseforge_to_resource_info(res: CurseForgeSearchRes) -> OtherResourceSearchRes {
+  let list = res
+    .data
+    .into_iter()
+    .map(|p| OtherResourceInfo {
+      id: p.id.to_string(),
+      _type: cvt_class_id_to_type(p.class_id),
+      name: p.name,
+      description: p.summary,
+      icon_src: p.logo.unwrap_or_else(default_logo).url,
+      website_url: p.links.website_url,
+      tags: p.categories.iter().map(|c| c.name.clone()).collect(),
+      last_updated: p.date_modified,
+      downloads: p.download_count,
+      source: OtherResourceSource::CurseForge,
+    })
+    .collect();
+
+  OtherResourceSearchRes {
+    list,
+    total: res.pagination.total_count,
+    page: res.pagination.index / res.pagination.page_size,
+    page_size: res.pagination.page_size,
+  }
+}
+
+fn extract_versions_and_loaders(game_versions: &[String]) -> (Vec<String>, Vec<String>) {
+  let mut versions = Vec::new();
+  let mut loaders = Vec::new();
+
+  const ALLOWED_LOADERS: &[&str] = &[
+    "Forge", "Fabric", "NeoForge", "Vanilla", "Iris", "Canvas", "OptiFine",
+  ];
+
+  for v in game_versions {
+    if v.starts_with(|c: char| c.is_ascii_digit()) {
+      versions.push(v.clone());
+    } else if ALLOWED_LOADERS.contains(&v.as_str()) {
+      loaders.push(v.clone());
+    }
+  }
+
+  (versions, loaders)
+}
+
+pub fn map_curseforge_file_to_version_pack(
+  res: Vec<CurseForgeFileInfo>,
+) -> Vec<OtherResourceVersionPack> {
+  let mut version_packs = std::collections::HashMap::new();
+
+  for cf_file in res {
+    let (versions, loaders) = extract_versions_and_loaders(&cf_file.game_versions);
+
+    let versions = if versions.is_empty() {
+      vec!["".to_string()]
+    } else {
+      versions
+    };
+
+    let loaders = if loaders.is_empty() {
+      vec![""]
+    } else {
+      loaders.iter().map(|s| s.as_str()).collect::<Vec<_>>()
+    };
+
+    for version in &versions {
+      for loader in &loaders {
+        let file_info = map_curseforge_file_info_to_other_resource_file_info(
+          &cf_file,
+          if loader.is_empty() {
+            None
+          } else {
+            Some(loader.to_string())
+          },
+        );
+
+        version_packs
+          .entry(version.clone())
+          .or_insert_with(|| OtherResourceVersionPack {
+            name: version.clone(),
+            items: Vec::new(),
+          })
+          .items
+          .push(file_info);
+      }
+    }
+  }
+
+  let mut list: Vec<OtherResourceVersionPack> = version_packs.into_values().collect();
+  list.sort_by(version_pack_sort);
+
+  list
 }
