+++ src-tauri/src/resource/helpers/modrinth.rs
@@ -5,202 +5,23 @@ use crate::resource::models::{
   OtherResourceVersionPackQuery, ResourceError,
 };
 use hex;
-use serde::Deserialize;
 use sha1::{Digest, Sha1};
 use std::collections::HashMap;
 use std::fs;
 use tauri::{AppHandle, Manager};
 use tauri_plugin_http::reqwest;
 
-use super::sort::version_pack_sort;
-
-#[derive(Deserialize, Debug)]
-pub struct ModrinthProject {
-  pub project_id: String,
-  pub project_type: String,
-  pub slug: String,
-  pub title: String,
-  pub description: String,
-  pub display_categories: Vec<String>,
-  pub downloads: u32,
-  pub icon_url: String,
-  pub date_modified: String,
-}
-
-#[derive(Deserialize, Debug)]
-pub struct ModrinthGetProjectRes {
-  pub id: String,
-  pub project_type: String,
-  pub slug: String,
-  pub title: String,
-  pub description: String,
-  pub categories: Vec<String>,
-  pub downloads: u32,
-  pub icon_url: String,
-  pub updated: String,
-}
-
-#[derive(Deserialize, Debug)]
-pub struct ModrinthSearchRes {
-  pub hits: Vec<ModrinthProject>,
-  pub total_hits: u32,
-  pub offset: u32,
-  pub limit: u32,
-}
-
-#[derive(Deserialize, Debug)]
-pub struct ModrinthFileHash {
-  pub sha1: String,
-}
-
-#[derive(Deserialize, Debug)]
-pub struct ModrinthFileInfo {
-  pub url: String,
-  pub filename: String,
-  pub hashes: ModrinthFileHash,
-}
-
-#[derive(Deserialize, Debug)]
-pub struct ModrinthFileDependency {
-  pub project_id: String,
-  pub dependency_type: String,
-}
-
-#[derive(Deserialize, Debug)]
-pub struct ModrinthVersionPack {
-  pub project_id: String,
-  pub dependencies: Vec<ModrinthFileDependency>,
-  pub game_versions: Vec<String>,
-  pub loaders: Vec<String>,
-  pub name: String,
-  pub date_published: String,
-  pub downloads: u32,
-  pub version_type: String,
-  pub files: Vec<ModrinthFileInfo>,
-}
-
-pub fn map_modrinth_to_resource_info(res: ModrinthSearchRes) -> OtherResourceSearchRes {
-  let list = res
-    .hits
-    .into_iter()
-    .map(|p| OtherResourceInfo {
-      id: p.project_id,
-      _type: p.project_type,
-      name: p.title,
-      description: p.description,
-      icon_src: p.icon_url,
-      website_url: format!("https://modrinth.com/mod/{}", p.slug),
-      tags: p.display_categories,
-      last_updated: p.date_modified,
-      downloads: p.downloads,
-      source: OtherResourceSource::Modrinth,
-    })
-    .collect();
-
-  OtherResourceSearchRes {
-    list,
-    total: res.total_hits,
-    page: res.offset / res.limit,
-    page_size: res.limit,
-  }
-}
-
-pub fn map_modrinth_file_to_version_pack(
-  res: Vec<ModrinthVersionPack>,
-) -> Vec<OtherResourceVersionPack> {
-  let mut version_packs = std::collections::HashMap::new();
-
-  for version in res {
-    let game_versions = if version.game_versions.is_empty() {
-      vec!["".to_string()]
-    } else {
-      version.game_versions
-    };
-
-    const ALLOWED_LOADERS: &[&str] = &[
-      "forge",
-      "fabric",
-      "neoforge",
-      "vanilla",
-      "iris",
-      "canvas",
-      "optifine",
-      "minecraft",
-    ];
-
-    let loaders = if version.loaders.is_empty() {
-      vec!["".to_string()]
-    } else {
-      version
-        .loaders
-        .iter()
-        .filter(|loader| ALLOWED_LOADERS.contains(&loader.as_str()))
-        .cloned()
-        .collect::<Vec<_>>()
-    };
-
-    for game_version in &game_versions {
-      for loader in &loaders {
-        let file_infos = version
-          .files
-          .iter()
-          .map(|file| OtherResourceFileInfo {
-            resource_id: version.project_id.clone(),
-            name: version.name.clone(),
-            release_type: version.version_type.clone(),
-            downloads: version.downloads,
-            file_date: version.date_published.clone(),
-            download_url: file.url.clone(),
-            sha1: file.hashes.sha1.clone(),
-            file_name: file.filename.clone(),
-            dependencies: version
-              .dependencies
-              .iter()
-              .map(|dep| OtherResourceDependency {
-                resource_id: dep.project_id.clone(),
-                relation: dep.dependency_type.clone(),
-              })
-              .collect(),
-            loader: if loader.is_empty() || loader == "minecraft" {
-              None
-            } else {
-              match loader.as_str() {
-                "forge" => Some("Forge".to_string()),
-                "fabric" => Some("Fabric".to_string()),
-                "neoforge" => Some("NeoForge".to_string()),
-                "vanilla" => Some("Vanilla".to_string()),
-                "iris" => Some("Iris".to_string()),
-                "canvas" => Some("Canvas".to_string()),
-                "optifine" => Some("OptiFine".to_string()),
-                _ => Some(loader.clone()),
-              }
-            },
-          })
-          .collect::<Vec<_>>();
-
-        version_packs
-          .entry(game_version.clone())
-          .or_insert_with(|| OtherResourceVersionPack {
-            name: game_version.clone(),
-            items: Vec::new(),
-          })
-          .items
-          .extend(file_infos);
-      }
-    }
-  }
-
-  let mut list: Vec<OtherResourceVersionPack> = version_packs.into_values().collect();
-  list.sort_by(version_pack_sort);
-
-  list
-}
+use super::modrinth_misc::{
+  get_modrinth_api, map_modrinth_file_to_version_pack, map_modrinth_to_resource_info,
+  map_modrinth_version_pack_to_other_resource_file_info, ModrinthApiEndpoint,
+  ModrinthGetProjectRes, ModrinthSearchRes, ModrinthVersionPack,
+};
 
 pub async fn fetch_resource_list_by_name_modrinth(
   app: &AppHandle,
   query: &OtherResourceSearchQuery,
 ) -> SJMCLResult<OtherResourceSearchRes> {
-  let url = "https://api.modrinth.com/v2/search";
+  let url = get_modrinth_api(ModrinthApiEndpoint::Search, None)?;
 
   let OtherResourceSearchQuery {
     resource_type,
@@ -229,7 +50,7 @@ pub async fn fetch_resource_list_by_name_modrinth(
 
   let client = app.state::<reqwest::Client>();
 
-  if let Ok(response) = client.get(url).query(&params).send().await {
+  if let Ok(response) = client.get(&url).query(&params).send().await {
     if response.status().is_success() {
       match response.json::<ModrinthSearchRes>().await {
         Ok(results) => Ok(map_modrinth_to_resource_info(results)),
@@ -253,10 +74,7 @@ pub async fn fetch_resource_version_packs_modrinth(
     game_versions,
   } = query;
 
-  let url = format!(
-    "https://api.modrinth.com/v2/project/{}/version",
-    resource_id
-  );
+  let url = get_modrinth_api(ModrinthApiEndpoint::ProjectVersions, Some(resource_id))?;
 
   let mut params = HashMap::new();
   if mod_loader != "All" {
@@ -278,7 +96,7 @@ pub async fn fetch_resource_version_packs_modrinth(
   let client = app.state::<reqwest::Client>();
 
   let response = client
-    .get(url)
+    .get(&url)
     .query(&params)
     .send()
     .await
@@ -310,7 +128,7 @@ pub async fn fetch_remote_resource_by_local_modrinth(
   let mut params = HashMap::new();
   params.insert("algorithm", "sha1");
 
-  let url = format!("https://api.modrinth.com/v2/version_file/{}", hash_string);
+  let url = get_modrinth_api(ModrinthApiEndpoint::VersionFile, Some(&hash_string))?;
   let client = app.state::<reqwest::Client>();
 
   let response = client
@@ -336,36 +154,22 @@ pub async fn fetch_remote_resource_by_local_modrinth(
     .or_else(|| version_pack.files.first())
     .ok_or(ResourceError::ParseError)?;
 
-  Ok(OtherResourceFileInfo {
-    resource_id: version_pack.project_id.clone(),
-    name: version_pack.name,
-    release_type: version_pack.version_type,
-    downloads: version_pack.downloads,
-    file_date: version_pack.date_published,
-    download_url: file_info.url.clone(),
-    sha1: file_info.hashes.sha1.clone(),
-    file_name: file_info.filename.clone(),
-    dependencies: version_pack
-      .dependencies
-      .iter()
-      .map(|dep| OtherResourceDependency {
-        resource_id: dep.project_id.clone(),
-        relation: dep.dependency_type.clone(),
-      })
-      .collect(),
-    loader: if version_pack.loaders.is_empty() {
+  Ok(map_modrinth_version_pack_to_other_resource_file_info(
+    &version_pack,
+    file_info,
+    if version_pack.loaders.is_empty() {
       None
     } else {
       Some(version_pack.loaders[0].clone())
     },
-  })
+  ))
 }
 
 pub async fn fetch_remote_resource_by_id_modrinth(
   app: &AppHandle,
   resource_id: &String,
 ) -> SJMCLResult<OtherResourceInfo> {
-  let url = format!("https://api.modrinth.com/v2/project/{}", resource_id);
+  let url = get_modrinth_api(ModrinthApiEndpoint::Project, Some(resource_id))?;
   let client = app.state::<reqwest::Client>();
 
   let response = client
